refactor(llm): Deduplicate DeepSeek prompt encoders
Make a deepseek_common module that is the superset of both versions. Everything that can live in the common module lives there, including items currently used by only one version. Only items that cannot be unified, primarily the two public formatter structs and their version-specific render_message / encode_messages pipelines, stay in the per-version files.

This keeps the per-version files quite small. Most of the code ends up in _common, and most of that is shared.

Implements the plan from #9317. All the details are there.

Closes: #9317
