+++ lib/llm/src/preprocessor/prompt.rs
@@ -25,6 +25,7 @@ use std::sync::Arc;
 
 use crate::preprocessor::media::MediaDecoder;
 
+pub mod deepseek_common;
 pub mod deepseek_v32;
 pub mod deepseek_v4;
 mod template;
