+++ lib/llm/src/preprocessor/prompt/deepseek_v4.rs
@@ -10,271 +10,13 @@
 use anyhow::{Context, Result};
 use serde_json::Value as JsonValue;
 
-/// Special tokens for DeepSeek V4
-pub mod tokens {
-    pub const BOS: &str = "<｜begin▁of▁sentence｜>";
-    pub const EOS: &str = "<｜end▁of▁sentence｜>";
-    pub const THINKING_START: &str = "<think>";
-    pub const THINKING_END: &str = "</think>";
-    pub const DSML_TOKEN: &str = "｜DSML｜";
-    pub const USER_START: &str = "<｜User｜>";
-    pub const ASSISTANT_START: &str = "<｜Assistant｜>";
-    pub const LATEST_REMINDER: &str = "<｜latest_reminder｜>";
-
-    // Quick-instruction task tokens
-    pub const TASK_ACTION: &str = "<｜action｜>";
-    pub const TASK_QUERY: &str = "<｜query｜>";
-    pub const TASK_AUTHORITY: &str = "<｜authority｜>";
-    pub const TASK_DOMAIN: &str = "<｜domain｜>";
-    pub const TASK_TITLE: &str = "<｜title｜>";
-    pub const TASK_READ_URL: &str = "<｜read_url｜>";
-}
-
-const TOOL_CALLS_BLOCK_NAME: &str = "tool_calls";
-
-const RESPONSE_FORMAT_TEMPLATE: &str =
-    "## Response Format:\n\nYou MUST strictly adhere to the following schema to reply:\n{schema}";
-
-const TOOLS_TEMPLATE: &str = r#"## Tools
-
-You have access to a set of tools to help answer the user's question. You can invoke tools by writing a "<{dsml_token}tool_calls>" block like the following:
-
-<{dsml_token}tool_calls>
-<{dsml_token}invoke name="$TOOL_NAME">
-<{dsml_token}parameter name="$PARAMETER_NAME" string="true|false">$PARAMETER_VALUE</{dsml_token}parameter>
-...
-</{dsml_token}invoke>
-<{dsml_token}invoke name="$TOOL_NAME2">
-...
-</{dsml_token}invoke>
-</{dsml_token}tool_calls>
-
-String parameters should be specified as is and set `string="true"`. For all other types (numbers, booleans, arrays, objects), pass the value in JSON format and set `string="false"`.
-
-If thinking_mode is enabled (triggered by {thinking_start_token}), you MUST output your complete reasoning inside {thinking_start_token}...{thinking_end_token} BEFORE any tool calls or final response.
-
-Otherwise, output directly after {thinking_end_token} with tool calls or final response.
-
-### Available Tool Schemas
-
-{tool_schemas}
-
-You MUST strictly follow the above defined tool name and parameter schemas to invoke tool calls.
-"#;
-
-const REASONING_EFFORT_MAX: &str = "Reasoning Effort: Absolute maximum with no shortcuts permitted.\nYou MUST be very thorough in your thinking and comprehensively decompose the problem to resolve the root cause, rigorously stress-testing your logic against all potential paths, edge cases, and adversarial scenarios.\nExplicitly write out your entire deliberation process, documenting every intermediate step, considered alternative, and rejected hypothesis to ensure absolutely no assumption is left unchecked.\n\n";
-
-/// Thinking mode for the model
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ThinkingMode {
-    Chat,
-    Thinking,
-}
-
-impl ThinkingMode {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            ThinkingMode::Chat => "chat",
-            ThinkingMode::Thinking => "thinking",
-        }
-    }
-}
-
-/// Reasoning effort level. `None` conveyed as `Option<ReasoningEffort>`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ReasoningEffort {
-    Max,
-    High,
-}
-
-/// Serialize a JSON value to match Python's `json.dumps(ensure_ascii=False)` spacing.
-/// Python's default separators are `(', ', ': ')`; we use a custom `Formatter`
-/// so escape sequences inside strings can't confuse state tracking.
-fn to_json(value: &JsonValue) -> String {
-    use serde::Serialize;
-    use serde_json::ser::Formatter;
-    use std::io;
-
-    struct PythonFormatter;
-
-    impl Formatter for PythonFormatter {
-        fn begin_array_value<W: ?Sized + io::Write>(
-            &mut self,
-            writer: &mut W,
-            first: bool,
-        ) -> io::Result<()> {
-            if first {
-                Ok(())
-            } else {
-                writer.write_all(b", ")
-            }
-        }
-
-        fn begin_object_key<W: ?Sized + io::Write>(
-            &mut self,
-            writer: &mut W,
-            first: bool,
-        ) -> io::Result<()> {
-            if first {
-                Ok(())
-            } else {
-                writer.write_all(b", ")
-            }
-        }
-
-        fn begin_object_value<W: ?Sized + io::Write>(&mut self, writer: &mut W) -> io::Result<()> {
-            writer.write_all(b": ")
-        }
-    }
-
-    let mut buf = Vec::with_capacity(64);
-    let mut ser = serde_json::Serializer::with_formatter(&mut buf, PythonFormatter);
-    if let Err(e) = value.serialize(&mut ser) {
-        tracing::warn!(error = %e, "to_json: serialize failed; falling back to empty object");
-        return "{}".to_string();
-    }
-    String::from_utf8(buf).unwrap_or_else(|e| {
-        tracing::warn!(error = %e, "to_json: serialized output not valid UTF-8; falling back to empty object");
-        "{}".to_string()
-    })
-}
-
-/// Extract function definitions from OpenAI-format tool list.
-fn tools_from_openai_format(tools: &[JsonValue]) -> Vec<JsonValue> {
-    tools
-        .iter()
-        .filter_map(|tool| tool.get("function").cloned())
-        .collect()
-}
-
-/// Render tool schemas into the system prompt format.
-fn render_tools(tools: &[JsonValue]) -> String {
-    let tools_json: Vec<String> = tools_from_openai_format(tools)
-        .iter()
-        .map(to_json)
-        .collect();
-
-    TOOLS_TEMPLATE
-        .replace("{tool_schemas}", &tools_json.join("\n"))
-        .replace("{dsml_token}", tokens::DSML_TOKEN)
-        .replace("{thinking_start_token}", tokens::THINKING_START)
-        .replace("{thinking_end_token}", tokens::THINKING_END)
-}
-
-/// Find the index of the last user/developer message.
-///
-/// Returns `None` when no such message exists. Callers should treat `None`
-/// as Python's `-1` sentinel: `idx >= -1` is always true in Python, so use
-/// `Option::is_none_or(|u| idx >= u)` (or `>`) to match the reference encoder.
-fn find_last_user_index(messages: &[JsonValue]) -> Option<usize> {
-    messages
-        .iter()
-        .enumerate()
-        .rev()
-        .find(|(_, msg)| {
-            msg.get("role")
-                .and_then(|r| r.as_str())
-                .map(|r| r == "user" || r == "developer")
-                .unwrap_or(false)
-        })
-        .map(|(idx, _)| idx)
-}
-
-/// Extract visible text from OpenAI-style message content.
-fn extract_visible_text(content: &JsonValue) -> String {
-    match content {
-        JsonValue::String(text) => text.clone(),
-        JsonValue::Array(items) => items
-            .iter()
-            .filter_map(|item| {
-                if let Some(text) = item.as_str() {
-                    return Some(text.to_string());
-                }
-                let item_type = item.get("type").and_then(|v| v.as_str());
-                if item_type == Some("text") {
-                    return item
-                        .get("text")
-                        .and_then(|v| v.as_str())
-                        .map(|text| text.to_string());
-                }
-                tracing::warn!(
-                    chunk_type = item_type.unwrap_or("unknown"),
-                    "DeepSeek V4 formatter dropped non-text content chunk while normalizing message content",
-                );
-                None
-            })
-            .collect::<String>(),
-        _ => to_json(content),
-    }
-}
-
-/// Normalize message `content` fields for text-only DeepSeek V4 rendering.
-fn normalize_message_contents(messages: &mut [JsonValue]) {
-    for msg in messages {
-        let Some(content) = msg.get("content") else {
-            continue;
-        };
-        // Leave non-string/non-array content untouched (null, etc.)
-        if !content.is_string() && !content.is_array() {
-            continue;
-        }
-        let normalized = extract_visible_text(content);
-        if let Some(obj) = msg.as_object_mut() {
-            obj.insert("content".to_string(), JsonValue::String(normalized));
-        }
-    }
-}
-
-/// Encode tool call arguments into DSML parameter format.
-fn encode_arguments_to_dsml(tool_call: &JsonValue) -> Result<String> {
-    let arguments_str = tool_call
-        .get("arguments")
-        .and_then(|a| a.as_str())
-        .context("Missing or invalid 'arguments' field")?;
-
-    // Python falls back to `{"arguments": raw_string}` on parse failure.
-    let arguments: JsonValue = match serde_json::from_str(arguments_str) {
-        Ok(v) => v,
-        Err(_) => serde_json::json!({ "arguments": arguments_str }),
-    };
-
-    let arguments_obj = arguments
-        .as_object()
-        .context("Arguments must be a JSON object")?;
-
-    let mut params = Vec::new();
-    for (key, value) in arguments_obj {
-        let is_string = value.is_string();
-        let value_str = if is_string {
-            value.as_str().unwrap().to_string()
-        } else {
-            to_json(value)
-        };
-        params.push(format!(
-            "<{}parameter name=\"{}\" string=\"{}\">{}</{}parameter>",
-            tokens::DSML_TOKEN,
-            key,
-            if is_string { "true" } else { "false" },
-            value_str,
-            tokens::DSML_TOKEN
-        ));
-    }
-
-    Ok(params.join("\n"))
-}
-
-/// Lookup the task token for a quick-instruction task.
-fn task_token(task: &str) -> Option<&'static str> {
-    match task {
-        "action" => Some(tokens::TASK_ACTION),
-        "query" => Some(tokens::TASK_QUERY),
-        "authority" => Some(tokens::TASK_AUTHORITY),
-        "domain" => Some(tokens::TASK_DOMAIN),
-        "title" => Some(tokens::TASK_TITLE),
-        "read_url" => Some(tokens::TASK_READ_URL),
-        _ => None,
-    }
-}
+use super::deepseek_common::{
+    NormalizeNonText, REASONING_EFFORT_MAX, RESPONSE_FORMAT_TEMPLATE, TOOL_CALLS_BLOCK_NAME,
+    TOOLS_TEMPLATE, drop_thinking_messages, encode_arguments_to_dsml, find_last_user_index,
+    merge_tool_messages, normalize_message_contents, render_tools, sort_tool_results_by_call_order,
+    task_token, to_json,
+};
+pub use super::deepseek_common::{ReasoningEffort, ThinkingMode, tokens};
 
 /// Render a single message at the given index.
 fn render_message(
@@ -308,7 +50,7 @@ fn render_message(
             prompt.push_str(content);
             if let Some(tools) = msg.get("tools").and_then(|t| t.as_array()) {
                 prompt.push_str("\n\n");
-                prompt.push_str(&render_tools(tools));
+                prompt.push_str(&render_tools(TOOLS_TEMPLATE, tools));
             }
             if let Some(response_format) = msg.get("response_format") {
                 prompt.push_str("\n\n");
@@ -330,7 +72,7 @@ fn render_message(
 
             if let Some(tools) = msg.get("tools").and_then(|t| t.as_array()) {
                 content_developer.push_str("\n\n");
-                content_developer.push_str(&render_tools(tools));
+                content_developer.push_str(&render_tools(TOOLS_TEMPLATE, tools));
             }
             if let Some(response_format) = msg.get("response_format") {
                 content_developer.push_str("\n\n");
@@ -515,182 +257,6 @@ fn render_tool_result_content(content: &JsonValue) -> String {
     }
 }
 
-/// Merge `tool` role messages into preceding user `content_blocks` and collapse
-/// consecutive user turns, matching Python's `merge_tool_messages`.
-pub fn merge_tool_messages(messages: &[JsonValue]) -> Vec<JsonValue> {
-    let mut merged: Vec<JsonValue> = Vec::with_capacity(messages.len());
-
-    for msg in messages {
-        let msg = msg.clone();
-        let role = msg.get("role").and_then(|r| r.as_str()).unwrap_or("");
-
-        if role == "tool" {
-            let tool_block = serde_json::json!({
-                "type": "tool_result",
-                "tool_use_id": msg.get("tool_call_id").cloned().unwrap_or(JsonValue::String(String::new())),
-                "content": msg.get("content").cloned().unwrap_or(JsonValue::String(String::new())),
-            });
-
-            let can_merge = merged
-                .last()
-                .map(|m| {
-                    m.get("role").and_then(|r| r.as_str()) == Some("user")
-                        && m.get("content_blocks").is_some()
-                })
-                .unwrap_or(false);
-
-            if can_merge {
-                let last = merged.last_mut().unwrap();
-                if let Some(blocks) = last
-                    .as_object_mut()
-                    .and_then(|o| o.get_mut("content_blocks"))
-                    .and_then(|v| v.as_array_mut())
-                {
-                    blocks.push(tool_block);
-                }
-            } else {
-                merged.push(serde_json::json!({
-                    "role": "user",
-                    "content_blocks": [tool_block],
-                }));
-            }
-        } else if role == "user" {
-            let text = msg
-                .get("content")
-                .and_then(|c| c.as_str())
-                .unwrap_or("")
-                .to_string();
-            let text_block = serde_json::json!({ "type": "text", "text": text });
-
-            let can_merge = merged
-                .last()
-                .map(|m| {
-                    m.get("role").and_then(|r| r.as_str()) == Some("user")
-                        && m.get("content_blocks").is_some()
-                        && m.get("task").map(|v| v.is_null()).unwrap_or(true)
-                })
-                .unwrap_or(false);
-
-            if can_merge {
-                let last = merged.last_mut().unwrap();
-                if let Some(blocks) = last
-                    .as_object_mut()
-                    .and_then(|o| o.get_mut("content_blocks"))
-                    .and_then(|v| v.as_array_mut())
-                {
-                    blocks.push(text_block);
-                }
-            } else {
-                let mut new_msg = serde_json::json!({
-                    "role": "user",
-                    "content": text,
-                    "content_blocks": [text_block],
-                });
-                // Preserve extra fields.
-                if let Some(obj) = new_msg.as_object_mut() {
-                    for key in ["task", "wo_eos", "mask"] {
-                        if let Some(v) = msg.get(key) {
-                            obj.insert(key.to_string(), v.clone());
-                        }
-                    }
-                }
-                merged.push(new_msg);
-            }
-        } else {
-            merged.push(msg);
-        }
-    }
-
-    merged
-}
-
-/// Sort `tool_result` blocks within user messages by the `tool_calls[].id` order
-/// of the preceding assistant message.
-pub fn sort_tool_results_by_call_order(mut messages: Vec<JsonValue>) -> Vec<JsonValue> {
-    use std::collections::HashMap;
-    let mut last_order: HashMap<String, usize> = HashMap::new();
-
-    for msg in &mut messages {
-        let role = msg.get("role").and_then(|r| r.as_str()).unwrap_or("");
-        if role == "assistant" {
-            if let Some(tcs) = msg.get("tool_calls").and_then(|t| t.as_array()) {
-                last_order.clear();
-                for (idx, tc) in tcs.iter().enumerate() {
-                    let id = tc
-                        .get("id")
-                        .and_then(|v| v.as_str())
-                        .or_else(|| {
-                            tc.get("function")
-                                .and_then(|f| f.get("id"))
-                                .and_then(|v| v.as_str())
-                        })
-                        .unwrap_or("");
-                    if !id.is_empty() {
-                        last_order.insert(id.to_string(), idx);
-                    }
-                }
-            }
-        } else if role == "user" && !last_order.is_empty() {
-            let Some(blocks) = msg
-                .as_object_mut()
-                .and_then(|o| o.get_mut("content_blocks"))
-                .and_then(|v| v.as_array_mut())
-            else {
-                continue;
-            };
-
-            // Collect tool_result blocks with their positions.
-            let tool_positions: Vec<usize> = blocks
-                .iter()
-                .enumerate()
-                .filter(|(_, b)| b.get("type").and_then(|v| v.as_str()) == Some("tool_result"))
-                .map(|(i, _)| i)
-                .collect();
-
-            if tool_positions.len() > 1 {
-                let mut tool_blocks: Vec<JsonValue> =
-                    tool_positions.iter().map(|&i| blocks[i].clone()).collect();
-                tool_blocks.sort_by_key(|b| {
-                    let id = b.get("tool_use_id").and_then(|v| v.as_str()).unwrap_or("");
-                    *last_order.get(id).unwrap_or(&0)
-                });
-                for (sorted_idx, &pos) in tool_positions.iter().enumerate() {
-                    blocks[pos] = tool_blocks[sorted_idx].clone();
-                }
-            }
-        }
-    }
-
-    messages
-}
-
-/// Drop reasoning and non-essential messages before the last user message.
-fn drop_thinking_messages(messages: Vec<JsonValue>) -> Vec<JsonValue> {
-    let last_user_idx = find_last_user_index(&messages);
-    let mut out = Vec::with_capacity(messages.len());
-    const KEEP: &[&str] = &[
-        "user",
-        "system",
-        "tool",
-        "latest_reminder",
-        "direct_search_results",
-    ];
-
-    for (idx, mut msg) in messages.into_iter().enumerate() {
-        let role = msg.get("role").and_then(|r| r.as_str()).unwrap_or("");
-        if KEEP.contains(&role) || last_user_idx.is_none_or(|u| idx >= u) {
-            out.push(msg);
-        } else if role == "assistant" {
-            if let Some(obj) = msg.as_object_mut() {
-                obj.remove("reasoning_content");
-            }
-            out.push(msg);
-        }
-        // developer and other roles before last_user_idx are dropped.
-    }
-    out
-}
-
 /// Encode messages to prompt string with default options.
 ///
 /// Equivalent to `encode_messages_with_options(.., drop_thinking=true, reasoning_effort=None)`.
@@ -857,7 +423,7 @@ impl super::OAIPromptFormatter for DeepSeekV4Formatter {
             .context("Messages is not an array")?
             .clone();
 
-        normalize_message_contents(&mut messages_array);
+        normalize_message_contents(&mut messages_array, NormalizeNonText::LeaveUntouched);
 
         let tools_json = req
             .tools()
@@ -1128,6 +694,29 @@ mod tests {
         ) -> Option<&std::collections::HashMap<String, serde_json::Value>> {
             self.chat_template_args.as_ref()
         }
+    }
+
+    #[test]
+    fn test_render_leaves_null_assistant_tool_content_empty() {
+        use super::super::OAIPromptFormatter;
+
+        let req = MockRequest::new(json!([
+            {"role": "user", "content": "call tool"},
+            {"role": "assistant", "content": null, "tool_calls": [{
+                "id": "c1", "type": "function",
+                "function": {"name": "f", "arguments": "{}"}
+            }]}
+        ]));
+
+        let formatter = DeepSeekV4Formatter::new_chat();
+        let out = formatter.render(&req).unwrap();
+
+        assert!(out.contains(&format!(
+            "<{}{}>",
+            tokens::DSML_TOKEN,
+            TOOL_CALLS_BLOCK_NAME
+        )));
+        assert!(!out.contains("null"));
     }
 
     #[test]
